Use SummaryCard on staking
